Rename yesod-lens stuff
- Consistently un-prefixed record field + "L"
- Consistnetly {module}.Lens
