+++ src/Yesod/Core/Types/Lens.hs
@@ -1,5 +1,5 @@
-module Yesod.Lens
-    ( handlerEnvL
+module Yesod.Core.Types.Lens
+    ( envL
     , siteL
     ) where
 
@@ -7,8 +7,8 @@ import RIO
 
 import Yesod.Core.Types (HandlerData(..), RunHandlerEnv(..))
 
-handlerEnvL :: Lens' (HandlerData child site) (RunHandlerEnv child site)
-handlerEnvL = lens handlerEnv $ \x y -> x { handlerEnv = y }
+envL :: Lens' (HandlerData child site) (RunHandlerEnv child site)
+envL = lens handlerEnv $ \x y -> x { handlerEnv = y }
 
 siteL :: Lens' (RunHandlerEnv child site) site
 siteL = lens rheSite $ \x y -> x { rheSite = y }
